mask-benchmark-replication/requirements.txt
inspect-ai==0.3.201
inspect-evals @ git+https://github.com/UKGovernmentBEIS/inspect_evals.git@0b1a737cf1d4dbd32235f47bf30bd731873dfc27
openai==2.30.0
anthropic==0.86.0
google-genai==1.68.0
xai-sdk==1.9.1
huggingface-hub==1.2.0
